+++ apps/web-antd/src/components/dict/src/type.d.ts
@@ -1,0 +1,5 @@
+/**
+ * fallback的渲染
+ * 可返回 字符串/Vnode
+ */
+export type DictFallback = ((current: number | string) => VNode) | string;